🚀 Add new feature: VLAN creation functionality
Implemented the `create_vlan` method in the `VLANConfig` class to allow users to create VLANs. The method accepts VLAN ID and VLAN name as input and returns the corresponding VLAN configuration. This feature enhances the capabilities of the VLAN subpackage and provides users with the ability to easily create VLANs.